pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aitodo"
version = "0.1.0"
authors = [
  { name="tsomoore", email="your_email@example.com" },
]
description = "An AI-powered, human-centric to-do list in your terminal."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer[all]",
    "rich",
    "google-generativeai",
]

[project.scripts]
aitodo = "aitodo.main:app"
